+++ src/components/loadable-list/index.js
@@ -1,0 +1,1 @@
+export { default as useLoadableList } from './use-loadable-list'
